+++ upcloud/load_balancer.go
@@ -252,9 +252,7 @@ type LoadBalancer struct {
 	Name             string                       `json:"name,omitempty"`
 	Zone             string                       `json:"zone,omitempty"`
 	Plan             string                       `json:"plan,omitempty"`
-	NetworkUUID      string                       `json:"network_uuid,omitempty"` // deprecated
 	Networks         []LoadBalancerNetwork        `json:"networks,omitempty"`
-	DNSName          string                       `json:"dns_name,omitempty"` // deprecated
 	Labels           []Label                      `json:"labels,omitempty"`
 	ConfiguredStatus LoadBalancerConfiguredStatus `json:"configured_status,omitempty"`
 	OperationalState LoadBalancerOperationalState `json:"operational_state,omitempty"`